initial stack logic
