+++ app/component/itinerary/NaviTop.js
@@ -5,11 +5,140 @@ import { itineraryShape, legShape } from '../../util/shapes';
 import { legTime, legTimeStr } from '../../util/legUtils';
 import NaviLeg from './NaviLeg';
 import Icon from '../Icon';
-import NaviStack from './NaviStack';
+import NaviStack from './NaviStack'; // milliseconds
+import { timeStr } from '../../util/timeUtils';
 
-function NaviTop({ itinerary, focusToLeg, time, realTimeLegs }) {
+const TRANSFER_SLACK = 60000;
+
+function findTransferProblem(legs) {
+  for (let i = 1; i < legs.length - 1; i++) {
+    const prev = legs[i - 1];
+    const leg = legs[i];
+    const next = legs[i + 1];
+
+    if (prev.transitLeg && leg.transitLeg && !leg.interlineWithPreviousLeg) {
+      // transfer at a stop
+      if (legTime(leg.start) - legTime(prev.end) < TRANSFER_SLACK) {
+        return [prev, leg];
+      }
+    }
+
+    if (prev.transitLeg && next.transitLeg && !leg.transitLeg) {
+      // transfer with some walking
+      const t1 = legTime(prev.end);
+      const t2 = legTime(next.start);
+      const transferDuration = legTime(leg.end) - legTime(leg.start);
+      const slack = t2 - t1 - transferDuration;
+      if (slack < TRANSFER_SLACK) {
+        return [prev, next];
+      }
+    }
+  }
+  return null;
+}
+const getScheduleInfo = (nextLeg, intl) => {
+  const { start, realTime, to, mode } = nextLeg;
+  const { estimatedTime, scheduledTime } = start;
+  const { parentStation, name } = to.stop;
+  const stack = [];
+
+  const localizedMode = intl.formatMessage({
+    id: `${mode.toLowerCase()}`,
+    defaultMessage: `${mode}`,
+  });
+
+  if (!realTime) {
+    stack.push({
+      content: (
+        <div className="navi-info-content">
+          <FormattedMessage id="navileg-mode-schedule" />
+          <FormattedMessage
+            id="navileg-start-schedule"
+            values={{
+              time: timeStr(scheduledTime),
+              mode: localizedMode,
+            }}
+          />{' '}
+        </div>
+      ),
+      backgroundColor: '#FFF8E8',
+      iconColor: '#FED100',
+      iconId: 'icon-icon_info',
+    });
+  } else {
+    const stopOrStation = parentStation
+      ? intl.formatMessage({ id: 'from-station' })
+      : intl.formatMessage({ id: 'from-stop' });
+
+    stack.push({
+      content: (
+        <div className="navi-info-content">
+          <FormattedMessage id="navileg-mode-realtime" values={{ mode }} />
+          <FormattedMessage
+            id="navileg-start-realtime"
+            values={{
+              time: timeStr(estimatedTime),
+              stopOrStation,
+              stopName: name,
+            }}
+          />
+        </div>
+      ),
+      backgroundColor: '##0074BF',
+      iconColor: '#FFF',
+      iconId: 'icon-icon_info',
+    });
+  }
+  return stack;
+};
+
+// We'll need the intl later. Thats certain.
+// eslint-disable-next-line no-unused-vars
+const getDisruptionInfo = (realTimeLegs, intl) => {
+  const stack = [];
+  const canceled = realTimeLegs.filter(leg => leg.realtimeState === 'CANCELED');
+  const transferProblem = findTransferProblem(realTimeLegs);
+  const late = realTimeLegs.filter(leg => leg.start.estimate?.delay > 0);
+
+  if (canceled.length > 0) {
+    // Todo: No current design
+    // todo find modes that are canceled
+    stack.push({
+      content: <div className="notifiler">Osa matkan lähdöistä on peruttu</div>,
+      backgroundColor: '#DC0451',
+      iconColor: '#888',
+      iconId: 'icon-icon_caution_white_exclamation',
+    });
+  }
+
+  if (transferProblem !== null) {
+    // todo no current design
+    stack.push({
+      content: (
+        <div className="notifiler">{`Vaihto ${transferProblem[0].route.shortName} - ${transferProblem[1].route.shortName} ei onnistu reittisuunnitelman mukaisesti`}</div>
+      ),
+      iconId: 'icon-icon_caution_white_exclamation',
+      iconColor: '#888',
+    });
+  }
+  if (late.length) {
+    // Todo: No current design
+    // Todo add mode and delay time to this message
+    stack.push({
+      content: <div className="notifiler">Kulkuneuvo on myöhässä</div>,
+      backgroundColor: '#FDF3F6',
+      iconColor: '#DC0451',
+      iconId: 'icon-icon_delay',
+    });
+  }
+
+  return stack;
+};
+
+function NaviTop({ itinerary, focusToLeg, time, realTimeLegs }, { intl }) {
   const [currentLeg, setCurrentLeg] = useState(null);
   const [show, setShow] = useState(true);
+  const [notifications, setNotifications] = useState([]);
 
   const handleClick = () => {
     setShow(!show);
@@ -20,19 +149,29 @@ function NaviTop({ itinerary, focusToLeg, time, realTimeLegs }) {
       setShow(false);
     }, 5000);
 
-    return () => clearTimeout(timer); // Cleanup the timer on component unmount
+    return () => clearTimeout(timer);
   }, []);
 
   useEffect(() => {
     const newLeg = realTimeLegs.find(leg => {
       return legTime(leg.start) <= time && time <= legTime(leg.end);
     });
-
+    const notis = [];
     if (newLeg?.id !== currentLeg?.id) {
       setCurrentLeg(newLeg);
       if (newLeg) {
         focusToLeg(newLeg, false);
+        notis.push(getScheduleInfo(newLeg, intl));
       }
+    }
+    const problems = getDisruptionInfo(realTimeLegs, intl);
+    if (problems.length > 0) {
+      notis.push(problems);
+    }
+    if (notis.length > 0) {
+      const combinedNotifications = notifications.concat(...notis);
+      setNotifications(combinedNotifications);
+      setShow(true);
     }
   }, [time]);
 
@@ -62,23 +201,31 @@ function NaviTop({ itinerary, focusToLeg, time, realTimeLegs }) {
   } else {
     info = <FormattedMessage id="navigation-wait" />;
   }
+  const handleRemove = index => {
+    setNotifications(notifications.filter((_, i) => i !== index));
+  };
 
   return (
     <>
       <button type="button" className="navitop" onClick={handleClick}>
         <div className="info">{info}</div>
         <div type="button" className="navitop-arrow">
-          {nextLeg && (
+          {nextLeg && notifications.length > 0 && (
             <Icon
               img="icon-icon_arrow-collapse"
               className={`cursor-pointer ${show ? 'inverted' : ''}`}
-              color="white"
+              color="blue"
             />
           )}
         </div>
       </button>
       {nextLeg && (
-        <NaviStack nextLeg={nextLeg} show={show} realTimeLegs={realTimeLegs} />
+        <NaviStack
+          stack={notifications}
+          setShow={setShow}
+          show={show}
+          handleRemove={handleRemove}
+        />
       )}
     </>
   );
