+++ app/component/itinerary/navigator.scss
@@ -52,7 +52,7 @@
   border-radius: 15px;
   height: 100px;
   color: white;
-  background-color: $primary-color !important;
+  background-color: white !important;
   display: flex;
   align-items: center;
 
@@ -78,19 +78,18 @@
       display: flex;
       flex-direction: row;
       height: 100%;
-      color: white;
+      color: black;
 
       .navileg-divider {
-        border: 1px solid #ffffff1a;
         height: 70px;
-        margin: 0 20px 0 20px;
+        margin: 0 15px 0 15px;
       }
 
       .navileg-mode {
         width: 40px;
         height: 40px;
         margin-top: 10px;
-        color: white;
+        color: black;
       }
     }
 
@@ -137,20 +136,24 @@
   }
 }
 
+@keyframes slideOutRight {
+  from {
+    transform: translateX(0);
+    opacity: 1;
+  }
+
+  to {
+    transform: translateX(-100%);
+    opacity: 0;
+  }
+}
+
 .info-stack {
   position: fixed;
   top: 180px;
   height: 69px;
   margin-left: 5%;
   width: 90%;
-  border-radius: 15px;
-  display: flex;
-  align-items: center;
-
-  .info-icon {
-    margin-top: 10px;
-    margin-left: 20px;
-  }
 
   &.slide-out {
     animation: slideOut 0.5s ease-out forwards;
@@ -161,33 +164,55 @@
     animation: slideIn 0.5s ease-out forwards;
   }
 
-  .navi-info {
+  .info-stack-item {
+    position: relative;
+    border-radius: 15px;
     display: flex;
-    flex-direction: column;
-    margin-left: 25px;
+    align-items: center;
+    height: 100%;
 
-    span:first-child {
-      font-weight: $font-weight-medium;
-      font-size: $font-size-normal;
+    &.slide-out-right {
+      animation: slideOutRight 0.5s ease-out forwards;
+      pointer-events: none;
     }
 
-    span:last-child {
-      font-size: $font-size-xsmall;
-    }
-  }
-
-  .navi-info-content {
-    display: flex;
-    flex-direction: column;
-    margin-left: 25px;
-
-    span:first-child {
-      font-weight: $font-weight-medium;
-      font-size: $font-size-normal;
+    .info-icon {
+      margin-top: 10px;
+      margin-left: 20px;
     }
 
-    span:last-child {
-      font-size: $font-size-xsmall;
+    &.slide-in {
+      animation: slideIn 0.5s ease-out forwards;
+    }
+
+    .navi-info {
+      display: flex;
+      flex-direction: column;
+      margin-left: 25px;
+
+      span:first-child {
+        font-weight: $font-weight-medium;
+        font-size: $font-size-normal;
+      }
+
+      span:last-child {
+        font-size: $font-size-xsmall;
+      }
+    }
+
+    .navi-info-content {
+      display: flex;
+      flex-direction: column;
+      margin-left: 25px;
+
+      span:first-child {
+        font-weight: $font-weight-medium;
+        font-size: $font-size-normal;
+      }
+
+      span:last-child {
+        font-size: $font-size-xsmall;
+      }
     }
   }
 }
